+++ modules/src/ics03_connection/handler/conn_open_ack.rs
@@ -1,13 +1,13 @@
 //! Protocol logic specific to processing ICS3 messages of type `MsgConnectionOpenAck`.
 
-use crate::events::IBCEvent;
+use crate::events::IbcEvent;
 use crate::handler::{HandlerOutput, HandlerResult};
 use crate::ics03_connection::connection::{ConnectionEnd, Counterparty, State};
 use crate::ics03_connection::context::ConnectionReader;
 use crate::ics03_connection::error::{Error, Kind};
 use crate::ics03_connection::events::Attributes;
 use crate::ics03_connection::handler::verify::{check_client_consensus_height, verify_proofs};
-use crate::ics03_connection::handler::ConnectionResult;
+use crate::ics03_connection::handler::{ConnectionIdState, ConnectionResult};
 use crate::ics03_connection::msgs::conn_open_ack::MsgConnectionOpenAck;
 
 pub(crate) fn process(
@@ -30,10 +30,15 @@ pub(crate) fn process(
                 || old_conn_end.state_matches(&State::TryOpen)
                     && old_conn_end.versions().get(0).eq(&Some(msg.version()));
 
-            // Check that if the msg's counterparty connection id is not empty then it matches
-            // the old connection's counterparty.
-            let counterparty_matches = msg.counterparty_connection_id().is_none()
-                || old_conn_end.counterparty().connection_id() == msg.counterparty_connection_id();
+            // Check that, if we have a counterparty connection id, then it
+            // matches the one in the message.
+            let counterparty_matches = if let Some(counterparty_connection_id) =
+                old_conn_end.counterparty().connection_id()
+            {
+                &msg.counterparty_connection_id == counterparty_connection_id
+            } else {
+                true
+            };
 
             if state_is_consistent && counterparty_matches {
                 Ok(old_conn_end)
@@ -59,8 +64,8 @@ pub(crate) fn process(
         Counterparty::new(
             // The counterparty is the local chain.
             new_conn_end.client_id().clone(), // The local client identifier.
-            msg.counterparty_connection_id().cloned(), // This chain's connection id as known on counterparty.
-            ctx.commitment_prefix(),                   // Local commitment prefix.
+            Some(msg.counterparty_connection_id().clone()), // This chain's connection id as known on counterparty.
+            ctx.commitment_prefix(),                        // Local commitment prefix.
         ),
         vec![msg.version().clone()],
         new_conn_end.delay_period,
@@ -81,15 +86,16 @@ pub(crate) fn process(
     new_conn_end.set_version(msg.version().clone());
 
     let result = ConnectionResult {
+        connection_id: msg.connection_id().clone(),
+        connection_id_state: ConnectionIdState::Reused,
         connection_end: new_conn_end,
-        connection_id: Some(msg.connection_id().clone()),
     };
 
     let event_attributes = Attributes {
-        connection_id: result.connection_id.clone(),
+        connection_id: Some(result.connection_id.clone()),
         ..Default::default()
     };
-    output.emit(IBCEvent::OpenAckConnection(event_attributes.into()));
+    output.emit(IbcEvent::OpenAckConnection(event_attributes.into()));
 
     Ok(output.with_result(result))
 }
@@ -99,18 +105,17 @@ mod tests {
     use std::convert::TryFrom;
     use std::str::FromStr;
 
-    use crate::events::IBCEvent;
+    use crate::events::IbcEvent;
     use crate::ics03_connection::connection::{ConnectionEnd, Counterparty, State};
     use crate::ics03_connection::error::Kind;
     use crate::ics03_connection::handler::{dispatch, ConnectionResult};
-    use crate::ics03_connection::msgs::conn_open_ack::test_util::get_dummy_msg_conn_open_ack;
+    use crate::ics03_connection::msgs::conn_open_ack::test_util::get_dummy_raw_msg_conn_open_ack;
     use crate::ics03_connection::msgs::conn_open_ack::MsgConnectionOpenAck;
     use crate::ics03_connection::msgs::ConnectionMsg;
     use crate::ics23_commitment::commitment::CommitmentPrefix;
     use crate::ics24_host::identifier::{ChainId, ClientId};
     use crate::mock::context::MockContext;
     use crate::mock::host::HostType;
-    use crate::Height;
 
     #[test]
     fn conn_open_ack_msg_processing() {
@@ -122,7 +127,8 @@ mod tests {
             error_kind: Option<Kind>,
         }
 
-        let msg_ack = MsgConnectionOpenAck::try_from(get_dummy_msg_conn_open_ack()).unwrap();
+        let msg_ack =
+            MsgConnectionOpenAck::try_from(get_dummy_raw_msg_conn_open_ack(10, 10)).unwrap();
         let conn_id = msg_ack.connection_id.clone();
 
         // Client parameters -- identifier and correct height (matching the proof height)
@@ -131,11 +137,13 @@ mod tests {
 
         // Parametrize the host chain to have a height at least as recent as the
         // the height of the proofs in the Ack msg.
+        let latest_height = proof_height.increment();
+        let max_history_size = 5;
         let default_context = MockContext::new(
-            ChainId::new("mockgaia".to_string(), 1),
+            ChainId::new("mockgaia".to_string(), latest_height.revision_number),
             HostType::Mock,
-            5,
-            Height::new(1, msg_ack.proofs().height().increment().revision_height),
+            max_history_size,
+            latest_height,
         );
 
         // A connection end that will exercise the successful path.
@@ -144,7 +152,7 @@ mod tests {
             client_id.clone(),
             Counterparty::new(
                 client_id.clone(),
-                msg_ack.counterparty_connection_id().cloned(),
+                Some(msg_ack.counterparty_connection_id().clone()),
                 CommitmentPrefix::from(b"ibc".to_vec()),
             ),
             vec![msg_ack.version().clone()],
@@ -161,18 +169,8 @@ mod tests {
         conn_end_prefix.set_state(State::Init);
         conn_end_prefix.set_counterparty(Counterparty::new(
             client_id.clone(),
-            msg_ack.counterparty_connection_id().cloned(),
+            Some(msg_ack.counterparty_connection_id().clone()),
             CommitmentPrefix::from(vec![]), // incorrect field
-        ));
-
-        // A connection end with correct state & prefix, but incorrect counterparty; exercises
-        // unsuccessful processing path.
-        let mut conn_end_cparty = conn_end_open.clone();
-        conn_end_cparty.set_state(State::Init);
-        conn_end_cparty.set_counterparty(Counterparty::new(
-            client_id.clone(),
-            None, // incorrect field
-            CommitmentPrefix::from(b"ibc".to_vec()),
         ));
 
         let tests: Vec<Test> = vec![
@@ -181,14 +179,14 @@ mod tests {
                 ctx: default_context
                     .clone()
                     .with_client(&client_id, proof_height)
-                    .with_connection(conn_id.clone(), default_conn_end.clone()),
+                    .with_connection(conn_id.clone(), default_conn_end),
                 msg: ConnectionMsg::ConnectionOpenAck(Box::new(msg_ack.clone())),
                 want_pass: true,
                 error_kind: None,
             },
             Test {
                 name: "Processing fails because the connection does not exist in the context".to_string(),
-                ctx: MockContext::default(),   // Empty context
+                ctx: default_context.clone(),
                 msg: ConnectionMsg::ConnectionOpenAck(Box::new(msg_ack.clone())),
                 want_pass: false,
                 error_kind: Some(Kind::UninitializedConnection(conn_id.clone())),
@@ -206,22 +204,13 @@ mod tests {
             Test {
                 name: "Processing fails: ConsensusStateVerificationFailure due to empty counterparty prefix".to_string(),
                 ctx: default_context
-                    .clone()
                     .with_client(&client_id, proof_height)
-                    .with_connection(conn_id.clone(), conn_end_prefix),
-                msg: ConnectionMsg::ConnectionOpenAck(Box::new(msg_ack.clone())),
+                    .with_connection(conn_id, conn_end_prefix),
+                msg: ConnectionMsg::ConnectionOpenAck(Box::new(msg_ack)),
                 want_pass: false,
                 error_kind: Some(Kind::ConsensusStateVerificationFailure(proof_height))
             },
-            Test {
-                name: "Processing fails due to mismatching counterparty conn id".to_string(),
-                ctx: default_context
-                    .with_client(&client_id, proof_height)
-                    .with_connection(conn_id.clone(), conn_end_cparty),
-                msg: ConnectionMsg::ConnectionOpenAck(Box::new(msg_ack.clone())),
-                want_pass: false,
-                error_kind: Some(Kind::ConnectionMismatch(conn_id.clone()))
-            },
+            /*
             Test {
                 name: "Processing fails due to MissingLocalConsensusState".to_string(),
                 ctx: MockContext::default()
@@ -231,6 +220,7 @@ mod tests {
                 want_pass: false,
                 error_kind: Some(Kind::MissingLocalConsensusState)
             },
+            */
         ];
 
         for test in tests {
@@ -253,11 +243,10 @@ mod tests {
                     assert_eq!(res.connection_end.state().clone(), State::Open);
 
                     for e in proto_output.events.iter() {
-                        assert!(matches!(e, &IBCEvent::OpenAckConnection(_)));
+                        assert!(matches!(e, &IbcEvent::OpenAckConnection(_)));
                     }
                 }
                 Err(e) => {
-                    println!("Error for {:?} was {:#?}", test.name, e.kind());
                     assert_eq!(
                         test.want_pass,
                         false,
@@ -272,7 +261,10 @@ mod tests {
                         assert_eq!(
                             &expected_kind,
                             e.kind(),
-                            "conn_open_ack: expected error kind mismatches thrown error kind"
+                            "conn_open_ack: failed for test: {}\nexpected error kind: {:?}\nfound: {:?}",
+                            test.name,
+                            expected_kind,
+                            e.kind()
                         )
                     }
                 }