+++ modules/src/ics04_channel/error.rs
@@ -3,9 +3,11 @@ use thiserror::Error;
 
 pub type Error = anomaly::Error<Kind>;
 
-use crate::ics24_host::identifier::{ChannelId, ConnectionId};
+use crate::ics04_channel::channel::State;
+use crate::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
+use crate::Height;
 
-#[derive(Clone, Debug, Error)]
+#[derive(Clone, Debug, Error, Eq, PartialEq)]
 pub enum Kind {
     #[error("channel state unknown")]
     UnknownState,
@@ -16,8 +18,8 @@ pub enum Kind {
     #[error("channel order type unknown")]
     UnknownOrderType,
 
-    #[error("invalid connection hops length")]
-    InvalidConnectionHopsLength,
+    #[error("invalid connection hops length: expected {0}; actual {1}")]
+    InvalidConnectionHopsLength(usize, usize),
 
     #[error("invalid version")]
     InvalidVersion,
@@ -55,8 +57,8 @@ pub enum Kind {
     #[error("given connection hop {0} does not exist")]
     MissingConnection(ConnectionId),
 
-    #[error("the port has no capability associated")]
-    NoPortCapability,
+    #[error("the port {0} has no capability associated")]
+    NoPortCapability(PortId),
 
     #[error("the module associated with the port does not have the capability it needs")]
     InvalidPortCapability,
@@ -67,8 +69,8 @@ pub enum Kind {
     #[error("the channel ordering is not supported by connection ")]
     ChannelFeatureNotSuportedByConnection,
 
-    #[error("queried for a non-existing connection")]
-    ChannelNotFound,
+    #[error("the channel end ({0}, {1}) does not exist")]
+    ChannelNotFound(PortId, ChannelId),
 
     #[error(
         "a different channel exists (was initialized) already for the same channel identifier {0}"
@@ -84,14 +86,14 @@ pub enum Kind {
     #[error("Channel chain verification fails on ChannelOpenTry for ChannelOpenInit")]
     FailedChanneOpenTryVerification,
 
-    #[error("No client state associated with the channel")]
-    MissingClientState,
+    #[error("No client state associated with client id {0}")]
+    MissingClientState(ClientId),
 
-    #[error("Frozen Client")]
-    FrozenClient,
+    #[error("Client with id {0} is frozen")]
+    FrozenClient(ClientId),
 
-    #[error("Missing client consensus state")]
-    MissingClientConsensusState,
+    #[error("Missing client consensus state for client id {0} at height {1}")]
+    MissingClientConsensusState(ClientId, Height),
 
     #[error("Invalid channel id in counterparty")]
     InvalidCounterpartyChannelId,
@@ -99,14 +101,17 @@ pub enum Kind {
     #[error("Client not found in chan open verification")]
     ClientNotFound,
 
-    #[error("Channel is in state {0} which is invalid")]
-    InvalidChannelState(ChannelId),
+    #[error("Channel {0} should not be state {1}")]
+    InvalidChannelState(ChannelId, State),
 
     #[error("Channel is in state {0}")]
     ChannelAlreadyClosed(ChannelId),
 
-    #[error("Channel chain verification fails on ChannelOpenAck for ChannelOpenTry")]
-    FailedChanneOpenAckVerification,
+    #[error("Handshake proof verification fails at ChannelOpenAck")]
+    ChanOpenAckProofVerification,
+
+    #[error("Handshake proof verification fails at ChannelOpenConfirm")]
+    ChanOpenConfirmProofVerification,
 }
 
 impl Kind {